fix(api): update secondary address parsing for types that require a number
The USPS says certain secondary addresses require a RANGE. At one point I assumed this was a number. Utah addresses don't always use a number and can sometimes be a range like APT L60-65 or something else UNIT A-SOUTH. The locators are smart enough to handle secondary addresses but since we don't include them in the build they lower the score. That shouldn't affect match rates much.

This code asumes a number or a single character and removes it. It's not perfect.